+++ src/JK.Common/CHANGELOG.md
@@ -12,10 +12,12 @@ Unless otherwise noted, all changes made by [@jeremyknight-me](https://github.co
 ### Added
 
 - Added `Slice` methods to `StringHelper` to take advantage of `Span<T>`
+- Added `DeserializeString` method to `XmlSerializationFacade`
 
 ### Changed
 
 - Changed Latitude and Longitude to use `decimal` instead of `double`
+- Renamed `GetXmlAsString` method to `Serialize` in `XmlSerializationFacade`
 
 ## 3.1.0